requirements.txt
Adafruit-Blinka==8.69.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-motor==3.4.18
adafruit-circuitpython-neopixel==6.3.21
adafruit-circuitpython-pca9685==3.4.20
adafruit-circuitpython-pixelbuf==2.0.10
adafruit-circuitpython-register==1.11.1
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-servokit==1.3.22
adafruit-circuitpython-typing==1.12.3
adafruit-circuitpython-veml7700==2.2.0
Adafruit-PlatformDetect==3.86.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
-e git+https://github.com/bitcraze/crazyflie-lib-python@99ad0e3e5be8ec717fd1b0fce0b7320e4acefe6e#egg=cflib
colorzero==2.0
gpiozero==2.0.1
importlib_resources==6.5.2
lgpio==0.2.2.0
libusb-package==1.0.26.3
numpy==2.4.1
opencv-python==4.13.0.92
packaging==25.0
phoenix6==25.4.2
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
rpi-lgpio==0.6
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
scipy==1.17.0
setuptools==80.9.0
smbus2==0.5.0
sysv_ipc==1.2.0
typing_extensions==4.15.0
